+++ src/card/sets/tests/runtime_support/nested_definitions.rs
@@ -376,6 +376,7 @@ pub(in super::super) fn assert_nested_definition_abilities(card_name: &str, effe
         | EffectDef::ReduceGenericCostBy(_)
         | EffectDef::LandwalkCanBeBlocked(_)
         | EffectDef::CannotAttackUnless(_)
+        | EffectDef::CannotAttackIf(_)
         | EffectDef::MoveToZone { .. }
         | EffectDef::Special(_) => {}
     }